refactor database util
