+++ libraries/common/src/main/java/androidx/media3/common/util/Util.java
@@ -47,8 +47,6 @@ import android.content.pm.PackageManager;
 import android.content.pm.PackageManager.NameNotFoundException;
 import android.content.res.Configuration;
 import android.content.res.Resources;
-import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteStatement;
 import android.graphics.Point;
 import android.graphics.drawable.Drawable;
 import android.hardware.display.DisplayManager;
@@ -85,8 +83,6 @@ import androidx.media3.common.ParserException;
 import androidx.media3.common.PlaybackException;
 import androidx.media3.common.Player;
 import androidx.media3.common.Player.Commands;
-import androidx.sqlite.db.SupportSQLiteDatabase;
-import androidx.sqlite.db.SupportSQLiteStatement;
 import com.google.common.base.Ascii;
 import com.google.common.base.Charsets;
 import com.google.common.util.concurrent.AsyncFunction;
@@ -2849,58 +2845,6 @@ public final class Util {
       removedItems.addFirst(items.remove(fromIndex + i));
     }
     items.addAll(min(newFromIndex, items.size()), removedItems);
-  }
-
-  /** Returns whether the table exists in the database. */
-  @UnstableApi
-  public static boolean tableExists(SupportSQLiteDatabase database, String tableName) {
-    long count =
-        queryNumEntries(database, "sqlite_master", "tbl_name = ?", new String[] {tableName});
-    return count > 0;
-  }
-
-  /**
-   * copied from the android framework {@link
-   * android.database.DatabaseUtils#queryNumEntries(SQLiteDatabase, String, String)}
-   */
-  public static long queryNumEntries(
-      SupportSQLiteDatabase db, String table, String selection, String[] selectionArgs) {
-    String s = (!android.text.TextUtils.isEmpty(selection)) ? " where " + selection : "";
-    return longForQuery(db, "select count(*) from " + table + s, selectionArgs);
-  }
-
-  /**
-   * Utility method to run the query on the db and return the value in the first column of the first
-   * row.
-   *
-   * <p>copied from {@link android.database.DatabaseUtils#longForQuery(SQLiteDatabase, String,
-   * String[])}
-   */
-  public static long longForQuery(SupportSQLiteDatabase db, String query, String[] selectionArgs) {
-    SupportSQLiteStatement prog = db.compileStatement(query);
-    try {
-      return longForQuery(prog, selectionArgs);
-    } finally {
-      androidx.media3.common.util.Util.closeQuietly(prog);
-    }
-  }
-
-  /** copied from {@link android.database.DatabaseUtils#longForQuery(SQLiteStatement, String[])} */
-  public static long longForQuery(SupportSQLiteStatement prog, String[] selectionArgs) {
-    bindAllArgsAsStrings(prog, selectionArgs);
-    return prog.simpleQueryForLong();
-  }
-
-  /**
-   * Equivalent to {@link android.database.sqlite.SQLiteStatement#bindAllArgsAsStrings(String[])}
-   */
-  public static void bindAllArgsAsStrings(
-      SupportSQLiteStatement prog, @Nullable String[] bindArgs) {
-    if (bindArgs != null) {
-      for (int i = bindArgs.length; i != 0; i--) {
-        prog.bindString(i, bindArgs[i - 1]);
-      }
-    }
   }
 
   /**
