+++ libraries/database/src/main/java/androidx/media3/database/VersionTable.java
@@ -23,7 +23,6 @@ import android.database.sqlite.SQLiteDatabase;
 import androidx.annotation.IntDef;
 import androidx.media3.common.MediaLibraryInfo;
 import androidx.media3.common.util.UnstableApi;
-import androidx.media3.common.util.Util;
 import androidx.sqlite.db.SupportSQLiteDatabase;
 import androidx.sqlite.db.SupportSQLiteQuery;
 import java.lang.annotation.Documented;
@@ -42,7 +41,7 @@ public final class VersionTable {
     MediaLibraryInfo.registerModule("media3.database");
   }
 
-  /** Returned by {@link #getVersion(SQLiteDatabase, int, String)} if the version is unset. */
+  /** Returned by {@link {@link VersionTable#getVersion(SupportSQLiteDatabase, int, String)} if the version is unset. */
   public static final int VERSION_UNSET = -1;
 
   /** Version of tables used for offline functionality. */
@@ -103,9 +102,8 @@ public final class VersionTable {
    * @param version The version.
    * @throws DatabaseIOException If an error occurs executing the SQL.
    */
-  public static void setVersion(SupportSQLiteDatabase writableDatabase, @Feature int feature,
-      String instanceUid,
-      int version)
+  public static void setVersion(
+      SupportSQLiteDatabase writableDatabase, @Feature int feature, String instanceUid, int version)
       throws DatabaseIOException {
     try {
       writableDatabase.execSQL(SQL_CREATE_TABLE_IF_NOT_EXISTS);
@@ -113,8 +111,7 @@ public final class VersionTable {
       values.put(COLUMN_FEATURE, feature);
       values.put(COLUMN_INSTANCE_UID, instanceUid);
       values.put(COLUMN_VERSION, version);
-      writableDatabase.insert(TABLE_NAME, SQLiteDatabase.CONFLICT_REPLACE,
-          values);
+      writableDatabase.insert(TABLE_NAME, SQLiteDatabase.CONFLICT_REPLACE, values);
     } catch (SQLException e) {
       throw new DatabaseIOException(e);
     }
@@ -132,7 +129,7 @@ public final class VersionTable {
       SupportSQLiteDatabase writableDatabase, @Feature int feature, String instanceUid)
       throws DatabaseIOException {
     try {
-      if (!Util.tableExists(writableDatabase, TABLE_NAME)) {
+      if (!DatabaseUtil.tableExists(writableDatabase, TABLE_NAME)) {
         return;
       }
       writableDatabase.delete(
@@ -154,18 +151,20 @@ public final class VersionTable {
    * @return The version, or {@link #VERSION_UNSET} if no version is set.
    * @throws DatabaseIOException If an error occurs executing the SQL.
    */
-  public static int getVersion(SupportSQLiteDatabase database, @Feature int feature,
-      String instanceUid)
+  public static int getVersion(
+      SupportSQLiteDatabase database, @Feature int feature, String instanceUid)
       throws DatabaseIOException {
     try {
-      if (!Util.tableExists(database, TABLE_NAME)) {
+      if (!DatabaseUtil.tableExists(database, TABLE_NAME)) {
         return VERSION_UNSET;
       }
-      SupportSQLiteQuery query = androidx.sqlite.db.SupportSQLiteQueryBuilder.builder(TABLE_NAME)
-          .columns(new String[]{COLUMN_VERSION})
-          .selection(WHERE_FEATURE_AND_INSTANCE_UID_EQUALS,
-              featureAndInstanceUidArguments(feature, instanceUid))
-          .create();
+      SupportSQLiteQuery query =
+          androidx.sqlite.db.SupportSQLiteQueryBuilder.builder(TABLE_NAME)
+              .columns(new String[] {COLUMN_VERSION})
+              .selection(
+                  WHERE_FEATURE_AND_INSTANCE_UID_EQUALS,
+                  featureAndInstanceUidArguments(feature, instanceUid))
+              .create();
       try (android.database.Cursor cursor = database.query(query)) {
         if (cursor.getCount() == 0) {
           return VERSION_UNSET;
@@ -179,6 +178,6 @@ public final class VersionTable {
   }
 
   private static String[] featureAndInstanceUidArguments(int feature, String instance) {
-    return new String[]{Integer.toString(feature), instance};
+    return new String[] {Integer.toString(feature), instance};
   }
 }
